Implement generic concept of recognizers with options pocketSphinx and phonetic
